// types/payload.ts
const LabelSchema = Type.Object({
  id: Type.Number(),
  node_id: Type.String(),
  url: Type.String(),
  name: Type.String(),
  color: Type.String(),
  default: Type.Boolean(),
  description: Type.Any(),
});
// ...
export enum UserType {
  User = "User",
  Bot = "Bot",
  Organization = "Organization",
}

export enum IssueType {
  OPEN = "open",
  CLOSED = "closed",
  // ALL = "all",
}

export enum StateReason {
  COMPLETED = "completed",
  NOT_PLANNED = "not_planned",
  REOPENED = "reopened",
}

const UserSchema = Type.Object({
  login: Type.String(),
  id: Type.Number(),
  node_id: Type.String(),
  avatar_url: Type.String(),
  gravatar_id: Type.Union([Type.Null(), Type.String()]),
  url: Type.String(),
  html_url: Type.String(),
  followers_url: Type.String(),
  following_url: Type.String(),
  gists_url: Type.String(),
  starred_url: Type.String(),
  subscriptions_url: Type.String(),
  organizations_url: Type.String(),
  repos_url: Type.String(),
  events_url: Type.String(),
  received_events_url: Type.String(),
  type: Type.Enum(UserType),
  site_admin: Type.Boolean(),
});
// ...
export type User = Static<typeof UserSchema>;
// type UserProfile= Static<typeof UserProfileSchema>;
export enum _AuthorAssociation {
  OWNER = "OWNER",
  COLLABORATOR = "COLLABORATOR",
  MEMBER = "MEMBER",
  CONTRIBUTOR = "CONTRIBUTOR",
  FIRST_TIMER = "FIRST_TIMER",
  FIRST_TIME_CONTRIBUTOR = "FIRST_TIME_CONTRIBUTOR",
  NONE = "NONE",
}
// const AuthorAssociation = Type.Enum(_AuthorAssociation);

const IssueSchema = Type.Object({
  assignee: Type.Union([Type.Null(), UserSchema]),
  assignees: Type.Array(Type.Union([Type.Null(), UserSchema])),
  author_association: Type.Enum(_AuthorAssociation),
  body: Type.String(),
  closed_at: Type.Union([Type.String({ format: "date-time" }), Type.Null()]),
  comments_url: Type.String(),
  comments: Type.Number(),
  created_at: Type.String({ format: "date-time" }),
  events_url: Type.String(),
  html_url: Type.String(),
  id: Type.Number(),
  labels_url: Type.String(),
  labels: Type.Array(LabelSchema),
  locked: Type.Boolean(),
  node_id: Type.String(),
  number: Type.Number(),
  repository_url: Type.String(),
  state_reason: Type.Union([Type.Enum(StateReason), Type.Null()]),
  state: Type.Enum(IssueType),
  title: Type.String(),
  updated_at: Type.String({ format: "date-time" }),
  url: Type.String(),
  user: UserSchema,
  // OWNER: The author is an owner of the repository.
  // COLLABORATOR: The author is a collaborator on the repository.
  // MEMBER: The author is a member of the organization that owns the repository.
  // CONTRIBUTOR: The author has contributed to the repository but is not a collaborator, member, or owner.
  // FIRST_TIMER: The author is a first-time contributor to the repository.
  // FIRST_TIME_CONTRIBUTOR: Similar to "FIRST_TIMER," the author is a first-time contributor to the repository.
  // NONE: The author does not have any specific association with the repository.
});

export type Issue = Static<typeof IssueSchema>;
// ...
const CommentSchema = Type.Object({
  author_association: Type.String(),
  body_html: Type.Optional(Type.String()),
  body_text: Type.Optional(Type.String()),
  body: Type.String(),
  created_at: Type.String({ format: "date-time" }),
  html_url: Type.String(),
  id: Type.Number(),
  issue_url: Type.String(),
  node_id: Type.String(),
  updated_at: Type.String({ format: "date-time" }),
  url: Type.String(),
  user: UserSchema,
  reactions: Type.Object({
    url: Type.String(),
    total_count: Type.Number(),
    "+1": Type.Number(),
    "-1": Type.Number(),
    laugh: Type.Number(),
    hooray: Type.Number(),
    confused: Type.Number(),
    heart: Type.Number(),
    rocket: Type.Number(),
    eyes: Type.Number(),
  }),
  // performed_via_github_app: Type.Optional(Type.Boolean()),
});

export type Comment = Static<typeof CommentSchema>;